Add new task function
