+++ src/index.html
@@ -13,6 +13,7 @@
       </div>
       <div class="data">
         <input id="data" type="text" placeholder="Add to your list...">
+        <i class="fas fa-level-down-alt"></i>
       </div>
       <div class="task-container">
         <ul id="sortList"></ul>
