+++ src/style.css
@@ -33,9 +33,8 @@ body {
 }
 
 #data {
-  width: -webkit-fill-available;
+  flex: 2;
   border: none;
-  border-top: 2px solid lightgrey;
   padding: 1rem;
 }
 
@@ -47,6 +46,18 @@ body {
 
 #data:focus {
   outline: none;
+}
+
+.data {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  border-top: 2px solid lightgrey;
+}
+
+.fa-level-down-alt {
+  padding: 15px;
+  color: lightgrey;
 }
 
 .drag {
